requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
black==21.6b0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
hexbytes==0.2.1
httplib2==0.19.1
idna==2.10
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
netaddr==0.8.0
oauthlib==3.1.1
packaging==21.0
parsimonious==0.8.1
pathspec==0.8.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.7.2
six==1.16.0
toml==0.10.2
toolz==0.11.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
varint==1.0.2
web3==5.20.1
websockets==9.1
yarl==1.6.3
